Add LOGSANITY option like LOGASSERT but disabled in release mode; fix ion::Matrix memory leaks (i.e. add constructors, destructors, assignment operator, etc.); make more ion::Matrix functions const; add ion::Matrix::AssertNotNan, ion::Matrix::View, ion::Matrix::Foreach (with arbitrary usrdata), ion::Matrix::operator- for left-hand constants, ion::Matrix::operator/ for left-hand constants, ion::Matrix::Countif, ion::Matrix::SwapRoi; disable default ion::Matrix construction completely (even for friends); fix bug in ion::Matrix::Swap; add ability to infer a dimension in ion::Matrix::Reshape; update AppFail to print error code; update logprintf to print time; add ion::Matrix<float> explicit instantiation; update ion::TimeGet to return time since program started instead of uptime; increase build warning level

+++ common/ionlib/src/net-win.cpp
@@ -94,7 +94,7 @@ namespace ion
 		send_to_addr.sin_addr.s_addr = address.as_integer();
 
 		int32_t bytes_sent = sendto(this->socket_handle_, buf, len, 0, (sockaddr*)&send_to_addr, sizeof(sockaddr_in));
-		if (bytes_sent != len)
+		if ((uint32_t)bytes_sent != len)
 		{
 			LOGERROR("Failed to send packet to %s:%hu, len: %u, error %d: %s", address.as_string(), port, GetLastError(), ion::getLastErrorString());
 			return ion::Error::Get(ion::Error::SOCKET);
